Improve sessiontalk and restructure (more class functions)
Move more of the code into a class to make the code easier to re-use for welcome emails.
Re-organize Default Settings to use sessiontalk group.
Misc bug fixes.
Improve instructions and windows softphone support
