+++ sessiontalk/app_config.php
@@ -46,64 +46,64 @@
 //default settings
 	$y=0;
 	$apps[$x]['default_settings'][$y]['default_setting_uuid'] = "67d5e1a7-d0b1-4f3a-b329-fabfb235b0c3";
-	$apps[$x]['default_settings'][$y]['default_setting_category'] = "provision";
-	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "sessiontalk_provider_id";
+	$apps[$x]['default_settings'][$y]['default_setting_category'] = "sessiontalk";
+	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "provider_id";
 	$apps[$x]['default_settings'][$y]['default_setting_name'] = "text";
 	$apps[$x]['default_settings'][$y]['default_setting_value'] = "";
 	$apps[$x]['default_settings'][$y]['default_setting_enabled'] = "true";
 	$apps[$x]['default_settings'][$y]['default_setting_description'] = "Sessiontalk Provider ID - Leave blank for whitelabel apps";
 	$y++;
 	$apps[$x]['default_settings'][$y]['default_setting_uuid'] = "f90a2cd2-55e3-42bf-a6b5-8c9db28ecc9b";
-	$apps[$x]['default_settings'][$y]['default_setting_category'] = "provision";
-	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "sessiontalk_transport";
+	$apps[$x]['default_settings'][$y]['default_setting_category'] = "sessiontalk";
+	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "transport";
 	$apps[$x]['default_settings'][$y]['default_setting_name'] = "text";
 	$apps[$x]['default_settings'][$y]['default_setting_value'] = "UDP";
 	$apps[$x]['default_settings'][$y]['default_setting_enabled'] = "false";
-	$apps[$x]['default_settings'][$y]['default_setting_description'] = "Sessiontalk Transport. Enable to force Transport type for Sessiontalk apps, otherwise uses the Transport from the Lines";
+	$apps[$x]['default_settings'][$y]['default_setting_description'] = "Sessiontalk Transport. Enable to set default Transport type for Sessiontalk devices.";
 	$y++;
 	$apps[$x]['default_settings'][$y]['default_setting_uuid'] = "432d9632-eafd-43ef-8955-3a510504c83f";
-	$apps[$x]['default_settings'][$y]['default_setting_category'] = "provision";
-	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "sessiontalk_srtp";
+	$apps[$x]['default_settings'][$y]['default_setting_category'] = "sessiontalk";
+	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "srtp";
 	$apps[$x]['default_settings'][$y]['default_setting_name'] = "text";
 	$apps[$x]['default_settings'][$y]['default_setting_value'] = "Disabled";
 	$apps[$x]['default_settings'][$y]['default_setting_enabled'] = "true";
 	$apps[$x]['default_settings'][$y]['default_setting_description'] = "Sessiontalk SRTP";
 	$y++;
 	$apps[$x]['default_settings'][$y]['default_setting_uuid'] = "1104ce84-a9fa-4b65-8376-c9990b40a41c";
-	$apps[$x]['default_settings'][$y]['default_setting_category'] = "provision";
-	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "sessiontalk_max_activations";
+	$apps[$x]['default_settings'][$y]['default_setting_category'] = "sessiontalk";
+	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "max_activations";
 	$apps[$x]['default_settings'][$y]['default_setting_name'] = "numeric";
 	$apps[$x]['default_settings'][$y]['default_setting_value'] = "1";
 	$apps[$x]['default_settings'][$y]['default_setting_enabled'] = "true";
 	$apps[$x]['default_settings'][$y]['default_setting_description'] = "Sessiontalk Maximum Apps per Extension. Defaults to 1 if not set, 0 for unlimited.";
 	$y++;		
 	$apps[$x]['default_settings'][$y]['default_setting_uuid'] = "6dbdeca0-131f-11eb-adc1-0242ac120002";
-	$apps[$x]['default_settings'][$y]['default_setting_category'] = "provision";
-	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "sessiontalk_qr_expiration";
+	$apps[$x]['default_settings'][$y]['default_setting_category'] = "sessiontalk";
+	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "qr_expiration";
 	$apps[$x]['default_settings'][$y]['default_setting_name'] = "numeric";
 	$apps[$x]['default_settings'][$y]['default_setting_value'] = "172800";
 	$apps[$x]['default_settings'][$y]['default_setting_enabled'] = "true";
 	$apps[$x]['default_settings'][$y]['default_setting_description'] = "QR Code is valid for this many seconds after being generated";
 	$y++;
 	$apps[$x]['default_settings'][$y]['default_setting_uuid'] = "14b71058-9361-43fd-9f0c-8a08876c7835";
-	$apps[$x]['default_settings'][$y]['default_setting_category'] = "provision";
-	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "sessiontalk_key_rotation";
+	$apps[$x]['default_settings'][$y]['default_setting_category'] = "sessiontalk";
+	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "key_rotation";
 	$apps[$x]['default_settings'][$y]['default_setting_name'] = "numeric";
-	$apps[$x]['default_settings'][$y]['default_setting_value'] = "1209600";
+	$apps[$x]['default_settings'][$y]['default_setting_value'] = "2592000";
 	$apps[$x]['default_settings'][$y]['default_setting_enabled'] = "true";
-	$apps[$x]['default_settings'][$y]['default_setting_description'] = "Rotation Period of keys. must be greater than the qr expiration period";
+	$apps[$x]['default_settings'][$y]['default_setting_description'] = "Rotation Period of keys. must be greater than the qr expiration period. If you reduce this you may invalidate some keys that have already been generated.";
 	$y++;
 	$apps[$x]['default_settings'][$y]['default_setting_uuid'] = "1af0c0f9-f19d-4f93-ad75-a5a5100d29c8";
-	$apps[$x]['default_settings'][$y]['default_setting_category'] = "provision";
-	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "sessiontalk_windows_softphone";
+	$apps[$x]['default_settings'][$y]['default_setting_category'] = "sessiontalk";
+	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "windows_softphone";
 	$apps[$x]['default_settings'][$y]['default_setting_name'] = "boolean";
 	$apps[$x]['default_settings'][$y]['default_setting_value'] = "true";
 	$apps[$x]['default_settings'][$y]['default_setting_enabled'] = "true";
 	$apps[$x]['default_settings'][$y]['default_setting_description'] = "Enable generating the Link to activate the Windows Softphone";
 	$y++;
 	$apps[$x]['default_settings'][$y]['default_setting_uuid'] = "8e38ad32-b124-4793-92a5-7f09478dfa51";
-	$apps[$x]['default_settings'][$y]['default_setting_category'] = "provision";
-	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "sessiontalk_windows_softphone_url";
+	$apps[$x]['default_settings'][$y]['default_setting_category'] = "sessiontalk";
+	$apps[$x]['default_settings'][$y]['default_setting_subcategory'] = "windows_softphone_url";
 	$apps[$x]['default_settings'][$y]['default_setting_name'] = "text";
 	$apps[$x]['default_settings'][$y]['default_setting_value'] = "https://windows-softphone.s3.eu-west-2.amazonaws.com/sessioncloud.appxbundle";
 	$apps[$x]['default_settings'][$y]['default_setting_enabled'] = "true";
